+++ src/components/Board/Board.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import {useState}from 'react';
 // import Navbar from './common/Navbar';
-import Sidebar from './common/Sidebar';
+import Sidebar from '../common/sidebar/Sidebar';
 import "./Board.css"
 
 export default function Board() {
