+++ src/components/Homepage/Homepage.js
@@ -6,12 +6,12 @@ import AuthContext from '../ContextApi/AuthContext';
 import Signup from '../LoginSignUp/Signup';
 import IntroPage from '../IntroPage/IntroPage';
 import LandingPage from '../LandingPage/LandingPage';
-import Navbar from '../common/navbar/Navbar'
-import Backlog from '../Backlog'
-import Roadmap from '../Roadmap';
-import Setting from '../Setting';
+import Navbar from '../common/navbar/Navbar'                                               
+import Backlog from '../Backlog/Backlog'
+import Roadmap from '../RoadMap/Roadmap';
+import Setting from '../Setting/Setting';
 import Project from '../common/navbar/Project';
-import Board from '../Board';
+import Board from '../Board/Board';
 
 const Homepage = () => {
   const { isAuthenticated,useEffectReload } = useContext(AuthContext);
