+++ src/components/Setting/Setting.js
@@ -1,7 +1,8 @@
 import React from 'react'
-import Sidebar from './common/Sidebar'
+// import Navbar from './common/Navbar'
+import Sidebar from '../common/sidebar/Sidebar'
 import { useState } from 'react';
-import "./Backlog.css"
+import "./Setting.css"
 
 export default function Backlog() {
   const [sidebarOpen, setSideBarOpen] = useState(true);
@@ -10,11 +11,11 @@ export default function Backlog() {
  }
 
   return (
-    <div className="backlog-component">
+    <div className="setting-component">
     {/* <Navbar onClick={handleViewSidebar} /> */}
     <Sidebar isOpen={sidebarOpen} toggleSidebar={handleViewSidebar} />
-    <div className="backlog-area">
-     <h1>Hello I am Backlog Page</h1>
+    <div className="setting-area">
+     <h1 className='heading'>Hello I am Setting Page</h1>
    </div>
    
  </div>
